requirements.txt
# Core Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# LangChain and LangGraph - Compatible versions
langchain>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.28

# Tokenizers (required by langchain-core)
transformers>=4.30.0
tokenizers>=0.13.0

# HTTP Requests
requests>=2.32.0

# Environment Variables
python-dotenv>=1.0.0
